requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
plotly>=5.13.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
anthropic>=0.15.0
matplotlib>=3.7.0
numpy>=1.24.0
openpyxl>=3.1.0
streamlit-calendar>=0.4.0